:sparkles: feat(frontend): 관광 아이템 선택 페이지에서만 보이는 오른쪽 사이드바 컴포넌트로 분리 (#287)
