+++ frontend/README.md
@@ -6,3 +6,5 @@
 - lottie requestAnimationFrame handler took 문제 해결 [해당이슈](https://github.com/YangGwangSeong/family-social/issues/59)
 - Animation 효과 적용
 - axios를 사용하여 rest-api를 호출할때 request data를 type-safe하게 사용하기 위해 axios.method 제네릭보다 `satisfies` 사용하기
+- 모바일 상태일때 줄였다 늘렸다 숨겼다 보였다 자유로운 UI를 가진 오른쪽 사이드 구현
+- 특정한 state 값에 따른 다른 페이지 랜더링을 효율적으로 랜더링 하는 방법에 대해
