+++ frontend/components/ui/layout/sidebar/schedule/sidebar-schedule-tourism/SidebarScheduleTourism.module.scss
@@ -1,0 +1,39 @@
+// 버튼 아래에 위치 시키기 위한 컨테이너
+.container {
+	@apply flex flex-col h-full;
+
+	.fix_container {
+		@apply relative;
+
+		.sidebar_tourism_total_time_container {
+			@apply flex my-4 gap-4 items-center;
+
+			.tourism_count {
+				@apply text-3xl font-semibold;
+			}
+
+			.stay_time {
+				@apply text-sm;
+			}
+		}
+	}
+
+	.right_sidebar_contents_wrap {
+		@apply flex-grow overflow-y-auto relative;
+
+		.schedule_tourism_container {
+			.not_found_tourism_container {
+				@apply flex justify-center items-center w-full h-[80px] bg-customOrange rounded-lg;
+
+				.not_found_text {
+					@apply w-full p-2 m-1 text-xs text-center text-customDark font-semibold;
+				}
+			}
+		}
+	}
+
+	// 버튼
+	.right_sidebar_footer_container {
+		@apply mb-2;
+	}
+}
